+++ pos_multi_session/pos_multi_session_models.py
@@ -7,6 +7,7 @@ class pos_config(models.Model):
 
     multi_session_id = fields.Many2one('pos.multi_session', 'Multi-session', help='Set the same value for POSes where orders should be synced. Keep empty if this POS should not use syncing')
     multi_session_accept_incoming_orders = fields.Boolean('Accept incoming orders', default=True)
+    multi_session_replace_empty_order = fields.Boolean('Replace empty order', default=True, help='Empty order is deleted whenever new order is come from another POS')
 
 
 class pos_multi_session(models.Model):