+++ pos_multi_session/doc/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 - Fix.Orders some times was out of sync. Now its ok.
 - Fix a bug related to updates in built-in bus module from Jan 20th 2016: https://github.com/odoo/odoo/commit/8af3841cb25cee33fd503ebe692abb8f98d4840a
 - Added demo data.
+- New: keep empty order. In previous version we deleted it when new Order from another POS is come. Now you can set it up in settings.
 
 
 `1.0.0`